睡眠グラフ表示用コンポーネントを作成
