+++ src/components/moods/SleepGraph.vue
@@ -1,28 +1,9 @@
 <template>
-    <ul class="tabnav">
-        <li @click="isActive = 'table'" :class="{ 'active': isActive === 'table' }">
-            睡眠表
-        </li>
-        <li @click="isActive = 'graph'" :class="{ 'active': isActive === 'graph' }">
-            睡眠グラフ
-        </li>
-    </ul>
-    <div v-if="isActive === 'table'">
-        <h1>睡眠表</h1>
-        <div class="border mb-5"></div>
-        <div class="mb-2">
-            <small class="me-3"><strong>※実睡眠時間</strong>：体が完全に安静な状態の時間</small>
-            <small class="me-3"><strong>※睡眠時間</strong>：ベッドにいた時間</small>
-            <small><strong>※睡眠効率</strong>：実睡眠時間 / 睡眠時間×100</small>
-        </div>
-    </div>
-    <div v-else-if="isActive === 'graph'">
-        <h1>睡眠グラフ</h1>
-        <div class="border mb-5"></div>
-        <ag-grid-vue class="ag-theme-alpine" style="height: 500px" :columnDefs="columnDefs" :rowData="rowData"
-            :defaultColDef="defaultColDef" rowSelection="multiple" animateRows="true">
-        </ag-grid-vue>
-    </div>
+    <h1>睡眠グラフ</h1>
+    <div class="border mb-5"></div>
+    <ag-grid-vue class="ag-theme-alpine" style="height: 500px" :columnDefs="columnDefs" :rowData="rowData"
+        :defaultColDef="defaultColDef" rowSelection="multiple" animateRows="true">
+    </ag-grid-vue>
 </template>
 
 <style>
@@ -34,25 +15,6 @@
 .ag-header-cell:not(:last-child) {
     border-right: solid 0.1px;
 }
-
-.tabnav {
-    display: flex;
-    list-style-type: none;
-    padding: 0;
-}
-
-.tabnav li {
-    cursor: pointer;
-    width: 50%;
-    background: #f2f3f5;
-    padding: 10px;
-    text-decoration: none;
-}
-
-.tabnav li.active {
-    background: #3cc0d9;
-    color: #fff;
-}
 </style>
 
 <script>
@@ -60,8 +22,6 @@ import { AgGridVue } from "ag-grid-vue3";  // the AG Grid Vue Component
 
 import "ag-grid-community/styles/ag-grid.css"; // Core grid CSS, always needed
 import "ag-grid-community/styles/ag-theme-alpine.css"; // Optional theme CSS
-
-import { ref } from 'vue';
 
 export default {
     name: "SleepGraph",
@@ -118,13 +78,10 @@ export default {
             flex: 1
         };
 
-        const isActive = ref('table')
-
         return {
             columnDefs,
             rowData,
-            defaultColDef,
-            isActive
+            defaultColDef
         };
     },
 };
